requirements.txt
google-generativeai
flask
pydantic
schedule
python-telegram-bot
chromadb
python-dotenv
requests
